+++ lib/Db/RecurrenceRule.php
@@ -32,12 +32,17 @@ use OCP\AppFramework\Db\Entity;
  * @method void setResetCheckboxes(bool $resetCheckboxes)
  * @method string getMode()
  * @method void setMode(string $mode)
+ * @method int getDueOffset()
+ * @method void setDueOffset(int $dueOffset)
  * @method int getCreatedAt()
  * @method void setCreatedAt(int $createdAt)
  */
 class RecurrenceRule extends Entity implements \JsonSerializable {
 	public const MODE_CLONE = 'clone';
 	public const MODE_RESET = 'reset';
+
+	/** dueOffset value meaning "spawned cards get no due date" */
+	public const DUE_NONE = -1;
 
 	protected string $userId = '';
 	protected int $templateCardId = 0;
@@ -50,6 +55,7 @@ class RecurrenceRule extends Entity implements \JsonSerializable {
 	protected bool $skipIfOpen = false;
 	protected bool $resetCheckboxes = false;
 	protected string $mode = self::MODE_CLONE;
+	protected int $dueOffset = 0;
 	protected int $createdAt = 0;
 
 	public function __construct() {
@@ -61,6 +67,7 @@ class RecurrenceRule extends Entity implements \JsonSerializable {
 		$this->addType('enabled', 'boolean');
 		$this->addType('skipIfOpen', 'boolean');
 		$this->addType('resetCheckboxes', 'boolean');
+		$this->addType('dueOffset', 'integer');
 		$this->addType('createdAt', 'integer');
 	}
 
@@ -77,6 +84,7 @@ class RecurrenceRule extends Entity implements \JsonSerializable {
 			'skipIfOpen' => $this->getSkipIfOpen(),
 			'resetCheckboxes' => $this->getResetCheckboxes(),
 			'mode' => $this->getMode(),
+			'dueOffset' => $this->getDueOffset(),
 		];
 	}
 }